requirements.txt
altair==4.2.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
decorator==5.1.1
deta==1.1.0
entrypoints==0.4
extra-streamlit-components==0.1.56
findspark==1.4.2
gitdb==4.0.10
GitPython==3.1.31
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hydralit-components==1.0.10
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.4.3
mdurl==0.1.2
numpy==1.21.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
plotly==5.3.1
premailer==3.10.0
protobuf==3.20.3
psutil==5.8.0
py4j==0.10.9.2
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.2.0
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
rfc3986==1.5.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
streamlit==1.22.0
streamlit-authenticator==0.2.1
streamlit-lottie==0.0.3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.0.4
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
yagmail==0.14.260
zipp==3.15.0
